Move the Explorer window root into nohrs-pages
Address review: the design must treat Explorer and the launcher as two
co-equal top-level pillars, and architecture.md §2 places per-feature views in
their crates while `nohrs-ui` owns window management (shared chrome).

`RootView` (the Explorer window's root: page navigation + search service) was
sitting in the binary, which baked an Explorer-only shape into the entry point.
Move it to `nohrs-pages` as the Explorer pillar — it depends only downward on
`nohrs-ui` and `nohrs-services` — so the future `nohrs-launcher` window root is
a symmetric sibling. Construction (page entities, progress loop) is folded into
`RootView::new`; the binary's `NohrsApp` is now a thin startup sequence that
initializes the search service (it owns the async runtime) and opens the window.
