+++ crates/nohrs/src/app.rs
@@ -1,32 +1,24 @@
-use nohrs_core::telemetry::logging::init_logging;
-use nohrs_pages::{
-    explorer::ExplorerPage, extensions::ExtensionsPage, git::GitPage, s3::S3Page,
-    settings::SettingsPage, PageKind,
-};
-use nohrs_ui::assets::Assets;
-use nohrs_ui::components::layout::footer::{footer, FooterProps};
-use nohrs_ui::components::layout::unified_toolbar::{
-    unified_toolbar, AccountMenuAction, AccountMenuCommand, UnifiedToolbarProps,
-    UNIFIED_TOOLBAR_HEIGHT,
-};
-use nohrs_ui::theme::theme;
-use nohrs_ui::window::{self, traffic_lights::TrafficLightsHook};
+//! Application entry / startup sequence.
+//!
+//! The binary is the only layer allowed to depend on every crate, so it wires
+//! the pillars together: it builds the shared window chrome from `nohrs-ui`,
+//! initializes services that need the async runtime, and opens the Explorer
+//! window hosting `nohrs_pages::RootView`. The launcher window (`nohrs-launcher`,
+//! P3) will be opened from here too, as a symmetric second pillar.
 
-use gpui::Entity;
-use gpui::{
-    div, prelude::*, px, rgb, size, AnyElement, App, AppContext, Application, Bounds, Context,
-    FocusHandle, Focusable, InteractiveElement, Render, Window,
-};
-use gpui_component::input::InputState;
+use gpui::{px, size, App, AppContext, Application, Bounds};
 use gpui_component::resizable::ResizableState;
-use gpui_component::Icon;
 use gpui_component::Root;
 use nohrs_core::config;
+use nohrs_core::telemetry::logging::init_logging;
+use nohrs_pages::RootView;
 use nohrs_services::search::SearchService;
+use nohrs_ui::assets::Assets;
+use nohrs_ui::components::layout::unified_toolbar::UNIFIED_TOOLBAR_HEIGHT;
+use nohrs_ui::window::{self, traffic_lights::TrafficLightsHook};
 use std::sync::Arc;
 use tokio::runtime::Handle;
 use tokio::task;
-use tracing::info;
 
 pub struct NohrsApp;
 
@@ -46,9 +38,6 @@ impl NohrsApp {
             let window_options = window::unified_window_options(bounds, &traffic_lights);
 
             let opened = app.open_window(window_options, |window, cx| {
-                let search_input = cx.new(|cx| InputState::new(window, cx));
-                let focus_handle = cx.focus_handle();
-
                 // Initialize SearchService. Failure is non-fatal: the app starts
                 // with full-text search disabled rather than crashing.
                 let handle = Handle::current();
@@ -64,259 +53,13 @@ impl NohrsApp {
                         }
                     });
 
-                // Create page instances
-                let explorer = cx.new(|cx| {
-                    ExplorerPage::new(
-                        resizable.clone(),
-                        search_input.clone(),
-                        search_service.clone(),
-                        cx.focus_handle(),
-                    )
-                });
-                // Search page removed
-                // let search = ...
-
-                let git = cx.new(|_cx| GitPage::new());
-                let s3 = cx.new(|_cx| S3Page::new());
-                let extensions = cx.new(|_cx| ExtensionsPage::new());
-                let settings = cx.new(|_cx| SettingsPage::new());
-
-                let view = cx.new(|cx| {
-                    let mut view = RootView {
-                        current_page: PageKind::Explorer,
-                        focus_handle,
-                        explorer,
-                        // search removed
-                        git,
-                        s3,
-                        extensions,
-                        settings,
-                        search_service: search_service.clone(),
-                        indexing_progress: Some(1.0), // Start as hidden/done
-                    };
-                    view.start_progress_loop(window, cx);
-                    view
-                });
-
+                let view =
+                    cx.new(|cx| RootView::new(resizable.clone(), search_service, window, cx));
                 cx.new(|cx| Root::new(view.into(), window, cx))
             });
             if let Err(error) = opened {
                 tracing::error!("failed to open main window: {error}");
             }
         });
-    }
-}
-
-pub struct RootView {
-    current_page: PageKind,
-    focus_handle: FocusHandle,
-    // Page entities
-    explorer: Entity<ExplorerPage>,
-    git: Entity<GitPage>,
-    s3: Entity<S3Page>,
-    extensions: Entity<ExtensionsPage>,
-    settings: Entity<SettingsPage>,
-    search_service: Option<Arc<SearchService>>,
-    indexing_progress: Option<f32>,
-}
-
-impl RootView {
-    pub fn set_page(&mut self, page: PageKind, cx: &mut Context<Self>) {
-        if self.current_page != page {
-            self.current_page = page;
-            cx.notify();
-        }
-    }
-
-    fn start_progress_loop(&mut self, window: &mut Window, cx: &mut Context<Self>) {
-        // Without a search service there is no indexing progress to poll, so avoid
-        // rescheduling a per-frame no-op forever.
-        if self.search_service.is_none() {
-            return;
-        }
-
-        if let Some(progress) = self.check_progress_update() {
-            self.indexing_progress = Some(progress);
-            cx.notify();
-        }
-
-        // Poll every frame (simple and effective for this case)
-        cx.on_next_frame(window, |view: &mut RootView, window, cx| {
-            view.start_progress_loop(window, cx);
-        });
-    }
-
-    fn check_progress_update(&self) -> Option<f32> {
-        let service = self.search_service.as_ref()?;
-        let rx = service.progress_subscription();
-        // Just read current value
-        let val = *rx.borrow();
-        Some(val)
-    }
-}
-
-impl Focusable for RootView {
-    fn focus_handle(&self, _cx: &App) -> FocusHandle {
-        self.focus_handle.clone()
-    }
-}
-
-impl Render for RootView {
-    fn render(&mut self, window: &mut Window, cx: &mut Context<Self>) -> impl IntoElement {
-        let toolbar = unified_toolbar(
-            UnifiedToolbarProps {
-                account_name: "syuya2036".to_string(),
-                account_plan: "Free".to_string(),
-            },
-            cx,
-        );
-
-        div()
-            .size_full()
-            .flex()
-            .flex_col()
-            .bg(rgb(theme::BG))
-            .relative()
-            .track_focus(&self.focus_handle)
-            .on_action(cx.listener(Self::handle_account_action))
-            .child(toolbar)
-            .child(
-                // Main content: toolbar + page
-                div()
-                    .flex_1()
-                    .flex()
-                    .flex_row()
-                    .min_h(px(0.0))
-                    .child(
-                        // Left navigation toolbar
-                        self.render_navigation(cx),
-                    )
-                    .child(
-                        // Main content area - render active page
-                        div()
-                            .flex_1()
-                            .flex()
-                            .flex_col()
-                            .min_w(px(0.0))
-                            .child(self.render_active_page(window, cx)),
-                    ),
-            )
-            .child(
-                // Footer status bar
-                {
-                    let (status_message, status_is_error) =
-                        match self.explorer.read(cx).status_for_footer() {
-                            Some((text, is_error)) => (Some(text), is_error),
-                            None => (None, false),
-                        };
-                    let props = FooterProps {
-                        indexing_progress: self.indexing_progress,
-                        status_message,
-                        status_is_error,
-                        ..Default::default()
-                    };
-                    footer(props, cx)
-                },
-            )
-            .children(Root::render_modal_layer(window, cx))
-            .children(Root::render_drawer_layer(window, cx))
-            .children(Root::render_notification_layer(window, cx))
-    }
-}
-
-impl RootView {
-    fn handle_account_action(
-        &mut self,
-        action: &AccountMenuAction,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        match action.command {
-            AccountMenuCommand::ProfileSummary => {
-                window.prevent_default();
-            }
-            AccountMenuCommand::Settings => self.set_page(PageKind::Settings, cx),
-            AccountMenuCommand::Extensions => self.set_page(PageKind::Extensions, cx),
-            AccountMenuCommand::Keymap
-            | AccountMenuCommand::Themes
-            | AccountMenuCommand::IconThemes => {
-                info!(?action.command, "Account menu item not yet implemented");
-                window.prevent_default();
-            }
-            AccountMenuCommand::SignOut => {
-                info!("Sign out requested");
-            }
-        }
-    }
-
-    fn render_navigation(&self, cx: &mut Context<Self>) -> impl IntoElement {
-        let active_page = self.current_page;
-
-        div()
-            .w(px(64.0))
-            .h_full()
-            .flex()
-            .flex_col()
-            .items_center()
-            .bg(rgb(theme::TOOLBAR_BG))
-            .border_r_1()
-            .border_color(rgb(theme::TOOLBAR_BORDER))
-            .py(px(16.0))
-            .child(
-                // Page navigation buttons
-                div().flex().flex_col().items_center().gap_2().children(
-                    PageKind::all().into_iter().map(|page| {
-                        let is_active = active_page == page;
-                        self.navigation_button(page, is_active, cx)
-                    }),
-                ),
-            )
-    }
-
-    fn navigation_button(
-        &self,
-        page: PageKind,
-        active: bool,
-        cx: &mut Context<Self>,
-    ) -> impl IntoElement {
-        div()
-            .id(("nav-btn", page as usize))
-            .w(px(48.0))
-            .h(px(48.0))
-            .flex()
-            .items_center()
-            .justify_center()
-            .rounded(px(8.0))
-            .cursor_pointer()
-            .when(active, |this| {
-                this.bg(rgb(theme::TOOLBAR_ACTIVE_BG)).shadow_sm()
-            })
-            .when(!active, |this| {
-                this.hover(|style| style.bg(rgb(theme::TOOLBAR_HOVER)))
-            })
-            .on_click(cx.listener(move |view, _event, _window, cx| {
-                view.set_page(page, cx);
-            }))
-            .child(
-                Icon::new(Icon::empty())
-                    .path(page.icon_path())
-                    .size_5()
-                    .text_color(rgb(if active {
-                        theme::TOOLBAR_ACTIVE_TEXT
-                    } else {
-                        theme::TOOLBAR_TEXT
-                    })),
-            )
-    }
-
-    fn render_active_page(&self, _window: &mut Window, _cx: &mut Context<Self>) -> AnyElement {
-        match self.current_page {
-            PageKind::Explorer => self.explorer.clone().into_any_element(),
-
-            PageKind::Git => self.git.clone().into_any_element(),
-            PageKind::S3 => self.s3.clone().into_any_element(),
-            PageKind::Extensions => self.extensions.clone().into_any_element(),
-            PageKind::Settings => self.settings.clone().into_any_element(),
-        }
     }
 }
